+++ transports/llm-direct-transport/src/LLMDirectTransport.ts
@@ -16,11 +16,11 @@ export interface LLMServiceOptions {
 }
 
 /**
- * RealTimeWebsocketTransport is an abstract class that provides a client-side
+ * LLMDirectTransport is an abstract class that provides a client-side
  * interface for connecting to a real-time AI service. It is intended to
  * connect directly to the service. (No Pipecat server is involved.)
  */
-export abstract class RealTimeWebsocketTransport extends Transport {
+export abstract class LLMDirectTransport extends Transport {
   // Utilities for audio.
   private _mediaManager;
   protected _service_options: LLMServiceOptions;